- Rename Params generic parameter to TModelIdentifier in RevealManagerBase - Gather types from utilities/networking/ in types.ts
