--- viewer/src/utilities/networking/CdfSource.ts
@@ -1,10 +1,0 @@
-/*!
- * Copyright 2020 Cognite AS
- */
-
-import { IdEither } from '@cognite/sdk';
-
-export interface CdfSource {
-  discriminator: 'cdf';
-  modelRevision: IdEither;
-}
