--- viewer/src/utilities/networking/ExternalSource.ts
@@ -1,8 +1,0 @@
-/*!
- * Copyright 2020 Cognite AS
- */
-
-export interface ExternalSource {
-  discriminator: 'external';
-  url: string;
-}
