--- viewer/src/utilities/networking/ModelUrlProvider.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright 2020 Cognite AS
- */
-
-export interface ModelUrlProvider<Params> {
-  getModelUrl(params: Params): Promise<string>;
-}
